[Regions] Remove region manager, add 'router'
The region manager was used as a means of integrating the router with
components on
a page. It was adding complexity and because it was instantiating
components the pathway
through 'component' binding handler was different meaning features such
as conventional
template names could not be determined.

A new binding handler,  'router' has been added which will listen for
route changes and pass through
a single 'component' named property through to a 'component' binding
handler.

Also, simplified the routing engine to be a bag of options for
configuration, instead of name, url and
options.